+++ docs/fastapi/http_codes.md
@@ -1,12 +1,12 @@
 # HTTP Codes Module
 
-Welcome to the `dsg_lib.fastapi_endpoints.http_codes` module. This module is a part of the `dsg_lib` library, designed to streamline and enhance your experience with the FastAPI framework.
+Welcome to the `dsg_lib.endpoints.http_codes` module. This module is a part of the `dsg_lib` library, designed to streamline and enhance your experience with the FastAPI framework.
 
 The `http_codes` module provides a set of constants representing HTTP status codes. These constants are designed to be used in FastAPI endpoints, making your code more readable and maintainable by replacing hard-coded status code numbers with meaningful constant names.
 
 Whether you are building a simple API or a complex web application, the `http_codes` module can help you ensure that your endpoints return the correct HTTP status codes. This not only improves the clarity of your code, but also helps to ensure that your API behaves in a way that is consistent with HTTP standards.
 
-In the following sections, we will guide you on how to import and use this module, along with examples to illustrate its usage in various scenarios. Whether you are a beginner or an experienced developer, we hope this documentation will be a valuable resource as you work with the `dsg_lib.fastapi_endpoints.http_codes` module.
+In the following sections, we will guide you on how to import and use this module, along with examples to illustrate its usage in various scenarios. Whether you are a beginner or an experienced developer, we hope this documentation will be a valuable resource as you work with the `dsg_lib.endpoints.http_codes` module.
 ## Using the Module
 
 The `http_codes` module provides constants for HTTP status codes. These constants can be used in FastAPI endpoints to set the HTTP status code of the response.
@@ -15,7 +15,7 @@ Here are examples of how to use the `http_codes` constants in FastAPI endpoints:
 
 ```python
 from fastapi import FastAPI
-from dsg_lib.fastapi_endpoints import http_codes
+from dsg_lib.endpoints import http_codes
 
 app = FastAPI()
