+++ docs/functions/logging.md
