+++ docs/functions/regex.md
