+++ dsg_lib/functions/patterns.py
