+++ tests/test_common_functions/test_file_functions/test_open_text.py
@@ -7,7 +7,7 @@ import time
 import unittest
 from pathlib import Path
 
-from dsg_lib.common_functions.file_functions import open_text
+from dsg_lib.functions.file_functions import open_text
 
 random_file_name_for_test = f"test_open_text_{secrets.token_hex(3)}.txt"
