+++ tests/test_common_functions/test_file_functions/test_save_csv.py
@@ -4,7 +4,7 @@ import unittest
 from pathlib import Path
 from unittest.mock import patch
 
-from dsg_lib.common_functions.file_functions import save_csv
+from dsg_lib.functions.file_functions import save_csv
 
 
 class TestFileFunctions(unittest.TestCase):
@@ -21,42 +21,42 @@ class TestFileFunctions(unittest.TestCase):
         if self.csv_path.exists():
             self.csv_path.unlink()
 
-    @patch("dsg_lib.common_functions.file_functions.directory_to_files", "data")
+    @patch("dsg_lib.functions.file_functions.directory_to_files", "data")
     def test_save_csv_with_valid_data(self):
         result = save_csv("test_file", self.test_data)
         self.assertEqual(result, "complete")
         self.assertTrue(self.csv_path.exists())
 
-    @patch("dsg_lib.common_functions.file_functions.directory_to_files", "data")
+    @patch("dsg_lib.functions.file_functions.directory_to_files", "data")
     def test_save_csv_with_invalid_data(self):
         with self.assertRaises(TypeError):
             save_csv("test_file", "not a list")
 
-    @patch("dsg_lib.common_functions.file_functions.directory_to_files", "data")
+    @patch("dsg_lib.functions.file_functions.directory_to_files", "data")
     def test_save_csv_with_invalid_file_name(self):
         with self.assertRaises(TypeError):
             save_csv("invalid/name", self.test_data)
 
-    @patch("dsg_lib.common_functions.file_functions.directory_to_files", "data")
+    @patch("dsg_lib.functions.file_functions.directory_to_files", "data")
     def test_save_csv_with_custom_delimiter(self):
         result = save_csv("test_file", self.test_data, delimiter=";")
         self.assertEqual(result, "complete")
         self.assertTrue(self.csv_path.exists())
 
-    @patch("dsg_lib.common_functions.file_functions.directory_to_files", "data")
+    @patch("dsg_lib.functions.file_functions.directory_to_files", "data")
     def test_save_csv_with_custom_quotechar(self):
         result = save_csv("test_file", self.test_data, quotechar="'")
         self.assertEqual(result, "complete")
         self.assertTrue(self.csv_path.exists())
 
-    @patch("dsg_lib.common_functions.file_functions.directory_to_files", "data")
+    @patch("dsg_lib.functions.file_functions.directory_to_files", "data")
     def test_save_csv_with_custom_root_folder(self):
         result = save_csv("test_file", self.test_data, root_folder="data/custom")
         self.assertEqual(result, "complete")
         custom_path = Path("data/custom/csv/test_file.csv")
         self.assertTrue(custom_path.exists())
 
-    @patch("dsg_lib.common_functions.file_functions.directory_to_files", "data")
+    @patch("dsg_lib.functions.file_functions.directory_to_files", "data")
     def test_save_csv_with_valid_data(self):
         result = save_csv("test_file", self.test_data)
         self.assertEqual(result, "complete")
