+++ tests/test_common_functions/test_file_functions/test_save_json.py
@@ -5,7 +5,7 @@ import unittest
 from pathlib import Path
 from typing import Dict, List
 
-from dsg_lib.common_functions.file_functions import save_json
+from dsg_lib.functions.file_functions import save_json
 
 
 class TestSaveJson(unittest.TestCase):
