+++ tests/test_common_functions/test_folder_functions/test_folder_functions_remove_folder.py
@@ -9,7 +9,7 @@ from pathlib import Path
 
 import pytest
 
-from dsg_lib.common_functions.folder_functions import remove_folder
+from dsg_lib.functions.folder_functions import remove_folder
 
 # from dsg_lib.folder_functions import get_directory_list
 # from dsg_lib.folder_functions import get_directory_list
