+++ tests/test_database_functions/test_http_codes.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from dsg_lib.fastapi_endpoints.http_codes import ALL_HTTP_CODES, generate_code_dict
+from dsg_lib.endpoints.http_codes import ALL_HTTP_CODES, generate_code_dict
 
 
 def test_generate_code_dict_description_only():
